Add remianing hooks + cleanups from PR
