+++ codegen/projections/high_score_service/lib/high_score_service/client.rb
@@ -420,7 +420,7 @@ module HighScoreService
     def initialize_config(config)
       config = config.dup
       client_interceptors = config.interceptors
-      config.interceptors = Hearth::Interceptor::List.new
+      config.interceptors = Hearth::InterceptorList.new
       Client.plugins.apply(config)
       Hearth::PluginList.new(config.plugins).apply(config)
       config.interceptors << client_interceptors
