+++ codegen/projections/rails_json/lib/rails_json/client.rb
@@ -2764,7 +2764,7 @@ module RailsJson
     def initialize_config(config)
       config = config.dup
       client_interceptors = config.interceptors
-      config.interceptors = Hearth::Interceptor::List.new
+      config.interceptors = Hearth::InterceptorList.new
       Client.plugins.apply(config)
       Hearth::PluginList.new(config.plugins).apply(config)
       config.interceptors << client_interceptors
