+++ codegen/projections/white_label/lib/white_label/client.rb
@@ -1016,7 +1016,7 @@ module WhiteLabel
     def initialize_config(config)
       config = config.dup
       client_interceptors = config.interceptors
-      config.interceptors = Hearth::Interceptor::List.new
+      config.interceptors = Hearth::InterceptorList.new
       Client.plugins.apply(config)
       Hearth::PluginList.new(config.plugins).apply(config)
       config.interceptors << client_interceptors
