+++ codegen/smithy-ruby-codegen-test/integration-specs/interceptor_spec.rb
@@ -9,16 +9,17 @@ module WhiteLabel
         def read_before_execution(context); end
       end.new
     end
+
     describe 'client configured interceptor' do
       it 'calls interceptor hook' do
         config = Config.new(
           stub_responses: true,
-          interceptors: Hearth::Interceptor::List.new([test_interceptor])
+          interceptors: Hearth::InterceptorList.new([test_interceptor])
         )
         client = Client.new(config)
 
         expect(test_interceptor).to receive(:read_before_execution)
-          .with(instance_of(Hearth::Interceptor::Context))
+          .with(instance_of(Hearth::InterceptorContext))
 
         client.kitchen_sink()
       end
