+++ hearth/.rubocop.yml
@@ -10,10 +10,13 @@ Metrics/AbcSize:
   Exclude:
     - 'spec/**/*.rb'
     - 'lib/hearth/http/client.rb'
+    - 'lib/hearth/middleware/retry.rb'
+    - 'lib/hearth/middleware/send.rb'
 
 Metrics/BlockLength:
   Exclude:
     - 'spec/**/*.rb'
+    - 'lib/hearth/middleware/retry.rb'
 
 Gemspec/RequireMFA:
   Exclude:
@@ -26,7 +29,7 @@ Metrics/MethodLength:
   Max: 20
   Exclude:
     - 'spec/**/*.rb'
-    - 'lib/hearth/middleware/send.rb'
+    - 'lib/hearth/middleware/*.rb'
 
 Metrics/ClassLength:
   Exclude:
